% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREETOPS.R
\name{finalize_TREETOPS}
\alias{finalize_TREETOPS}
\title{Getting final treetops}
\usage{
finalize_TREETOPS(sf_TREETOPS, distance, min_H, max_H = NULL)
}
\arguments{
\item{sf_TREETOPS}{sf object (output of 'get_TREETOPS()')}

\item{distance}{numeric, distance between treetops in meters}

\item{min_H}{numeric, minimum height of trees}

\item{max_H}{numeric, maximum height of trees, if not set (default ~ NULL)
the maximum height value of the input sf object (sf_TREETOPS)}
}
\value{
sf object
}
\description{
MAIN FUNCTION for distance based treetop filtering using the output of 'get_TREETOPS()'.
}
