% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREETOPS.R
\name{get_CCC}
\alias{get_CCC}
\title{Getting patches}
\usage{
get_CCC(level_raster, level_rasterMINone, level = NULL)
}
\arguments{
\item{level_raster}{/result of get_LEVELz()/ i-th layer of SpatRaster stack}

\item{level_rasterMINone}{/result of get_LEVELz()/ i - 1 layer of SpatRaster stack}

\item{level}{level layer, always level_rasterMINone, however the algorithm needs this extra parameter}
}
\value{
list (sf object and terra SpatRaster)
}
\description{
Function for getting the patches (clumps) and centroids of GTR, used in 'get_MARKER()' that is embedded in 'get_TREETOPS()'.
}
