% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREETOPS.R
\name{get_GTR}
\alias{get_GTR}
\title{Getting GTR and FETR}
\usage{
get_GTR(lmo_seg, new_tr)
}
\arguments{
\item{lmo_seg}{segregated patches defined in get_CCC, SpatRaster stack}

\item{new_tr}{new tree region, SpatRaster (value 1)}
}
\value{
SpatRaster (value 1 GTR pixels, value 0 non-GTR pixels)
}
\description{
Function for binarizing patches into GROWING TREE REGIONS (GTR ~ 1) and FIRST EMERGING TREE REGION (FETR ~ 0), used in 'get_CCC()'.
}
