% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREETOPS.R
\name{get_HB}
\alias{get_HB}
\title{Getting height levels I}
\usage{
get_HB(CHM_g, level_increment, min_H)
}
\arguments{
\item{CHM_g}{chm from las, SpatRaster}

\item{level_increment}{numeric, height bin}

\item{min_H}{numeric, under this value no trees are reasonable}
}
\value{
numeric vector
}
\description{
Function for getting height bin (numeric vector).
}
