% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREETOPS.R
\name{get_MARKER}
\alias{get_MARKER}
\title{Getting markers}
\usage{
get_MARKER(CHM_g, GTR_ccomponent, GTR_marker)
}
\arguments{
\item{CHM_g}{chm from las, SpatRaster}

\item{GTR_ccomponent}{/result of get_CCC/ GTR patches from layer2 to layer3, level raster is layer3, SpatRaster (values n patches)}

\item{GTR_marker}{/result of get_CCC/ GTR markers from layer1 to layer2, level raster is layer2, sf spatial points}
}
\value{
sf object
}
\description{
Function for getting markers, its ids and heights (Z), used in 'get_TREETOPS()'.
}
