% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREETOPS.R
\name{get_TREETOPS_optimized}
\alias{get_TREETOPS_optimized}
\title{Memory-optimized version of get_TREETOPS}
\usage{
get_TREETOPS_optimized(
  CHM_g,
  min_H,
  level_increment = 0.2,
  use_parallel = FALSE,
  n_cores = NULL,
  cores_tile = 1
)
}
\arguments{
\item{CHM_g}{SpatRaster object, raster chm derived from las}

\item{min_H}{numeric, minimum (Z) value of chm, higher values can be treetops}

\item{level_increment}{numeric, level cutting increment (default = 0.2m)}

\item{use_parallel}{logical, whether to use parallel processing (default = FALSE)}

\item{n_cores}{number of cores for parallel processing (default = NULL)}

\item{cores_tile}{number of tiles per core (default = 1)}
}
\value{
sf object
}
\description{
MAIN FUNCTION for obtaining treetops with memory optimization.
}
