% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREETOPS.R
\name{get_TREETOPS_parallel_controlled}
\alias{get_TREETOPS_parallel_controlled}
\title{Parallel processing with controlled tile distribution}
\usage{
get_TREETOPS_parallel_controlled(
  CHM_g,
  min_H,
  level_increment = 0.2,
  n_cores = NULL,
  cores_tile = 1
)
}
\arguments{
\item{CHM_g}{SpatRaster object, raster chm derived from las}

\item{min_H}{numeric, minimum (Z) value of chm}

\item{level_increment}{numeric, level cutting increment}

\item{n_cores}{number of parallel workers}

\item{cores_tile}{number of tiles per core (default = 1)}
}
\value{
sf object
}
\description{
Parallel processing with controlled tile distribution
}
