% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREETOPS.R
\name{M_diff_PROFILE}
\alias{M_diff_PROFILE}
\title{Helper function for obtaining metrics}
\usage{
M_diff_PROFILE(profile, Zstart, Zstop)
}
\arguments{
\item{profile}{~  height (Z) profile between reference tree and test treetops}

\item{Zstart}{height of reference tree}

\item{Zstop}{height of test treetop}
}
\value{
data frame
}
\description{
Used in 'M_match_TREETOPS()'.
}
