% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TREETOPS.R
\name{M_match_TREETOPS}
\alias{M_match_TREETOPS}
\title{Matching treetops}
\usage{
M_match_TREETOPS(refTT, testTT, CHM_g, eval_maxHD = 2, eval_maxD = 4)
}
\arguments{
\item{refTT}{sf object, reference trees}

\item{testTT}{sf object, either GTR or VWF results, extracted treetops}

\item{CHM_g}{SpatRAster, Gaussian smoothed chm generated from las}

\item{eval_maxHD}{numeric or double, minimum height difference (m) btw. reference and test trees (default = 2)}

\item{eval_maxD}{numeric or double, minimum distance btw. reference (m) and test trees (default = 4)}
}
\value{
data.table
}
\description{
The function finds the 3 nearest test tree points (knn), matches one (height difference and distance) and returns:
- height difference between observed and test tree,
- distance bewteen observed and test tree.
}
